extract BNF AST utilities from generator to BnfAst
isTrivialNode, getNonTrivialNode, getTrivialNodeChild, getEffectiveType,
compilePattern, and the textStrategy hash helper are pure BNF tree
operations — they have nothing to do with code generation. Move them to
a new top-level org.intellij.grammar.psi.BnfAst.

Continues prep for extracting :generator: BnfFileImpl, BnfStringImpl,
BnfFirstNextAnalyzer, GrammarUtil no longer touch ParserGeneratorUtil
for AST navigation/regex compilation.
