+++ bnf-language/src/org/intellij/grammar/generator/ExpressionGeneratorHelper.java
@@ -20,6 +20,7 @@ import java.util.*;
 
 import static java.lang.String.format;
 import static org.intellij.grammar.psi.BnfAttributes.*;
+import static org.intellij.grammar.psi.BnfAst.*;
 import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
 
 /**
