+++ bnf-language/src/org/intellij/grammar/generator/Generator.java
@@ -29,6 +29,7 @@ import java.util.*;
 
 import static java.lang.String.format;
 import static org.intellij.grammar.psi.BnfAttributes.*;
+import static org.intellij.grammar.psi.BnfAst.*;
 import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
 import static org.intellij.grammar.generator.RuleGraphHelper.*;
