+++ bnf-language/src/org/intellij/grammar/psi/BnfAst.java
@@ -1,0 +1,108 @@
+/*
+ * Copyright 2011-2025 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+ */
+package org.intellij.grammar.psi;
+
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.tree.IElementType;
+import com.intellij.psi.util.PsiTreeUtil;
+import it.unimi.dsi.fastutil.Hash;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.Objects;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
+
+import static org.intellij.grammar.psi.BnfTypes.BNF_SEQUENCE;
+
+public final class BnfAst {
+
+  private BnfAst() { }
+
+  private static final Hash.Strategy<PsiElement> TEXT_STRATEGY = new Hash.Strategy<>() {
+    @Override
+    public int hashCode(PsiElement e) {
+      return e == null ? 0 : e.getText().hashCode();
+    }
+
+    @Override
+    public boolean equals(PsiElement e1, PsiElement e2) {
+      return e1 == null ? e2 == null : e2 != null && Objects.equals(e1.getText(), e2.getText());
+    }
+  };
+
+  @SuppressWarnings("unchecked")
+  public static <T extends PsiElement> Hash.Strategy<T> textStrategy() {
+    return (Hash.Strategy<T>)TEXT_STRATEGY;
+  }
+
+  public static boolean isTrivialNode(PsiElement element) {
+    return getTrivialNodeChild(element) != null;
+  }
+
+  public static BnfExpression getNonTrivialNode(BnfExpression initialNode) {
+    BnfExpression nonTrivialNode = initialNode;
+    for (BnfExpression e = initialNode, n = getTrivialNodeChild(e); n != null; e = n, n = getTrivialNodeChild(e)) {
+      nonTrivialNode = n;
+    }
+    return nonTrivialNode;
+  }
+
+  public static BnfExpression getTrivialNodeChild(PsiElement element) {
+    PsiElement child = null;
+    if (element instanceof BnfParenthesized) {
+      BnfExpression e = ((BnfParenthesized)element).getExpression();
+      if (element instanceof BnfParenExpression) {
+        child = e;
+      }
+      else {
+        BnfExpression c = e;
+        while (c instanceof BnfParenthesized) {
+          c = ((BnfParenthesized)c).getExpression();
+        }
+        if (c.getFirstChild() == null) {
+          child = e;
+        }
+      }
+    }
+    else if (element.getFirstChild() == element.getLastChild() && element instanceof BnfExpression) {
+      child = element.getFirstChild();
+    }
+    return child instanceof BnfExpression && !(child instanceof BnfLiteralExpression || child instanceof BnfReferenceOrToken) ?
+           (BnfExpression)child : null;
+  }
+
+  public static IElementType getEffectiveType(PsiElement tree) {
+    if (tree instanceof BnfParenOptExpression) {
+      return BnfTypes.BNF_OP_OPT;
+    }
+    else if (tree instanceof BnfQuantified) {
+      BnfQuantifier quantifier = ((BnfQuantified)tree).getQuantifier();
+      return PsiTreeUtil.getDeepestFirst(quantifier).getNode().getElementType();
+    }
+    else if (tree instanceof BnfPredicate) {
+      return ((BnfPredicate)tree).getPredicateSign().getFirstChild().getNode().getElementType();
+    }
+    else if (tree instanceof BnfStringLiteralExpression) {
+      return BnfTypes.BNF_STRING;
+    }
+    else if (tree instanceof BnfLiteralExpression) {
+      return tree.getFirstChild().getNode().getElementType();
+    }
+    else if (tree instanceof BnfParenExpression) {
+      return BNF_SEQUENCE;
+    }
+    else {
+      return tree.getNode().getElementType();
+    }
+  }
+
+  public static @Nullable Pattern compilePattern(String text) {
+    try {
+      return Pattern.compile(text);
+    }
+    catch (PatternSyntaxException e) {
+      return null;
+    }
+  }
+}
