+++ bnf-language/src/org/intellij/grammar/psi/impl/GrammarUtil.java
@@ -31,6 +31,7 @@ import java.util.Collections;
 import java.util.List;
 
 import static com.intellij.psi.SyntaxTraverser.psiTraverser;
+import static org.intellij.grammar.psi.BnfAst.*;
 import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
 import static org.intellij.grammar.psi.BnfAttributes.*;
 import static org.intellij.grammar.psi.BnfTypes.BNF_SEQUENCE;
