+++ src/org/intellij/grammar/livePreview/LivePreviewParser.java
@@ -27,6 +27,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
+import static org.intellij.grammar.psi.BnfAst.*;
 import static org.intellij.grammar.generator.ParserGeneratorUtil.*;
 import static org.intellij.grammar.parser.GeneratedParserUtilBase.*;
 import static org.intellij.grammar.psi.BnfAttributes.*;
